# specs/004-intermediate-features/contracts/api.yaml
openapi: 3.0.3
info:
  title: Todo App - Intermediate Features API
  description: API specification for categories, priorities, tags, search, statistics, subtasks, and attachments
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api
    description: Local development
  - url: https://api.todoapp.com/api
    description: Production
security:
  - BearerAuth: []
paths:
...
  /tasks/{task_id}/tags:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer

    post:
      summary: Add tag to task
      tags: [Tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 30
      responses:
        '201':
          description: Tag added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
  /tasks/{task_id}/tags/{tag_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
      - name: tag_id
        in: path
        required: true
        schema:
          type: integer

    delete:
      summary: Remove tag from task
      tags: [Tags]
      responses:
        '204':
          description: Tag removed

  # ========================================
  # Attachments
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
...